requirements.txt
geopandas
numpy
pandas
pandas-datareader
pandas-profiling
pandas-summary
pandasql
scikit-learn
scikit-learn-intelex
sklearn-pandas
timezonefinder


django
pickle-mixin
pillow
requests
timezonefinder
geopy
pytz
folium
google-generativeai